backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
Pillow>=10.4.0
websockets>=12.0
pyngrok>=7.0.0
qrcode>=7.4.2
